co-locate all templating types
