+++ packages/insomnia/src/network/network.ts
@@ -33,7 +33,7 @@ import { isWorkspace, type Workspace } from '../models/workspace';
 import * as pluginContexts from '../plugins/context/index';
 import * as plugins from '../plugins/index';
 import { RenderError } from '../templating/render-error';
-import type { ExtraRenderInfo, RenderedRequest, RenderPurpose, RequestAndContext } from '../templating/types';
+import type { RenderedRequest, RenderPurpose, RequestAndContext } from '../templating/types';
 import { maskOrDecryptContextIfNecessary } from '../templating/utils';
 import { defaultSendActionRuntime, type SendActionRuntime } from '../ui/routes/request';
 import { invariant } from '../utils/invariant';
@@ -611,7 +611,7 @@ export const tryToInterpolateRequest = async ({
   request: Request;
   environment: string | Environment;
   purpose?: RenderPurpose;
-  extraInfo?: ExtraRenderInfo;
+  extraInfo?: { name: string; value: any }[];
   baseEnvironment?: Environment;
   userUploadEnvironment?: UserUploadEnvironment;
   transientVariables?: Environment;
