+++ packages/insomnia/src/templating/base-extension.ts
@@ -4,9 +4,9 @@ import type { Request } from '../models/request';
 import type { RequestGroup } from '../models/request-group';
 import type { Workspace } from '../models/workspace';
 import * as pluginContexts from '../plugins/context';
-import type { PluginTemplateTag } from './extensions';
+import type { Plugin } from '../plugins/index';
 import * as templating from './index';
-import type { HelperContext } from './types';
+import type { PluginTemplateTag, PluginTemplateTagContext } from './types';
 import { decodeEncoding } from './utils';
 
 const EMPTY_ARG = '__EMPTY_NUNJUCKS_ARG__';
@@ -63,9 +63,8 @@ export default class BaseExtension {
     return this._ext?.deprecated || false;
   }
 
-  run(...args: any[]) {
-    // @ts-expect-error -- TSCONVERSION
-    return this._ext?.run(...args);
+  run(context: PluginTemplateTagContext, ...arg: any[]) {
+    return this._ext?.run(context, ...arg);
   }
 
   parse(parser: any, nodes: any, lexer: any) {
@@ -97,7 +96,7 @@ export default class BaseExtension {
       .filter(a => a !== EMPTY_ARG)
       .map(decodeEncoding);
     // Define a helper context with utils
-    const helperContext: HelperContext = {
+    const helperContext: PluginTemplateTagContext = {
       ...pluginContexts.app.init(renderPurpose),
       // @ts-expect-error -- TSCONVERSION
       ...pluginContexts.store.init(this._plugin),
