+++ packages/insomnia/src/ui/components/codemirror/code-editor.tsx
@@ -16,7 +16,8 @@ import { DEBOUNCE_MILLIS, isMac } from '../../../common/constants';
 import * as misc from '../../../common/misc';
 import type { KeyCombination } from '../../../common/settings';
 import { getTagDefinitions } from '../../../templating/index';
-import { extractNunjucksTagFromCoords, type NunjucksParsedTag, type nunjucksTagContextMenuOptions } from '../../../templating/utils';
+import { type NunjucksParsedTag, type nunjucksTagContextMenuOptions } from '../../../templating/types';
+import { extractNunjucksTagFromCoords } from '../../../templating/utils';
 import { ednPrettify } from '../../../utils/prettify/edn';
 import { jsonPrettify } from '../../../utils/prettify/json';
 import { queryXPath } from '../../../utils/xpath/query';
